EL02 Miscellaneous fixes and corrections
Everything now compiles well and corrrectly

+++ Tools/Cosmology/ModeFunctions.h
@@ -15,8 +15,8 @@ using namespace std;
 
 typedef double (*modefunction)(double, Parameters &);
 
-static double plainFunction(int a, Parameters & /* plainParameters */){
-    return (double) a;
+static double plainFunction(double a, Parameters & /* plainParameters */){
+    return a;
 }
 
 struct method{
